Great refactoring 2:
- move dg scripts to scripts
- update rst files
- update documentation
- change burgess to burgers
- add example descriptions fo gallery
