+++ doc/developer_guide.rst
@@ -1085,10 +1085,13 @@ sfepy.discrete.dg sub-package
 .. toctree::
    :maxdepth: 2
 
-   src/sfepy/discrete/dg/dg_field
-   src/sfepy/discrete/dg/dg_basis
-   src/sfepy/discrete/dg/dg_terms
-   src/sfepy/discrete/dg/dg_limiters
+   src/sfepy/discrete/dg/fields
+   src/sfepy/discrete/dg/dg_poly_spaces
+   src/sfepy/discrete/dg/limiters
+   src/sfepy/discrete/dg/gen_legendre_simplex_base
+   src/script/dg_plot_1D
+   src/script/dg_1D_vizualizer
+   src/sfepy/solvers/ts_dg_solvers
 
 
 sfepy.discrete.iga sub-package
