+++ doc/src/script/dg_1D_vizualizer.rst
@@ -1,0 +1,6 @@
+script/dg_1D_vizualizer.py script
+=================================
+
+.. automodule:: dg_1D_vizualizer
+   :members:
+   :undoc-members:
