+++ doc/src/script/dg_plot_1D.rst
@@ -1,0 +1,6 @@
+script/dg_plot_1D.py script
+===========================
+
+.. automodule:: dg_plot_1D
+   :members:
+   :undoc-members:
