--- doc/src/sfepy/discrete/dg/dg_basis.rst
@@ -1,7 +1,0 @@
-sfepy.dicrete.dg.dg_basis module
-================================
-
-.. automodule:: sfepy.discrete.dg.dg_poly_space
-   :members:
-   :undoc-members:
-
