+++ doc/src/sfepy/discrete/dg/dg_poly_spaces.rst
@@ -1,0 +1,7 @@
+sfepy.dicrete.dg.dg_poly_spaces module
+================================
+
+.. automodule:: sfepy.discrete.dg.dg_poly_spaces
+   :members:
+   :undoc-members:
+
