--- doc/src/sfepy/discrete/dg/dg_terms.rst
@@ -1,6 +1,0 @@
-sfepy.dicrete.dg.dg_terms module
-================================
-
-.. automodule:: sfepy.terms.terms_dg
-   :members:
-   :undoc-members:
