+++ doc/src/sfepy/discrete/dg/gen_legendre_simplex_base.rst
@@ -1,0 +1,6 @@
+sfepy.discrete.dg.gen_legendre_simplex_base module
+==================================================
+
+.. automodule:: sfepy.discrete.dg.gen_legendre_simplex_base
+   :members:
+   :undoc-members:
