+++ doc/src/sfepy/solvers/ts_dg_solvers.rst
@@ -1,0 +1,6 @@
+sfepy.solvers.ts_dg_solvers module
+==================================
+
+.. automodule:: sfepy.solvers.ts_dg_solvers
+   :members:
+   :undoc-members:
