+++ examples/dg/inits_consts.py
@@ -1,8 +1,5 @@
 # -*- coding: utf-8 -*-
-"""
-Different initial conditions and environment constants as
-functions to enable different samplig in solvers
-"""
+""" Different initial conditions and environment constants as functions """
 import numpy as nm
 from scipy import signal
 
@@ -67,9 +64,6 @@ def three_step_q(x):
 def three_step_u(x):
     """
     piecewise constant (-inf, a],(a, a + 5](a+5, inf)
-
-    :param x:
-    :return:
     """
     return nm.piecewise(x, [x <= 0, x <= .7, .7 < x], [0, .5, 2])
 
@@ -77,9 +71,6 @@ def three_step_u(x):
 def four_step_u(x):
     """
     piecewise constant (-inf, 1.8],(1.8, a + 4](a+4, a + 5](a + 5, inf)
-
-    :param x:
-    :return:
     """
     return nm.piecewise(x, [x <= a, x <= a + .4, a + .4 < x, a + .5 <= x], [0, 0, .5, 0])
 
@@ -87,9 +78,6 @@ def four_step_u(x):
 def four_step_q(x):
     """
     piecewise constant (-inf, 1.8],(1.8, a + 4](a+4, a + 5](a + 5, inf)
-
-    :param x:
-    :return:
     """
     return nm.piecewise(x, [x <= a, x <= a + .4, a + .4 < x, a + .5 <= x], [0, 0, 1, 0])
 
@@ -99,8 +87,6 @@ def four_step_q(x):
 def const_u(x):
     """
     piecewise constant (-inf, a],(a, a + 5](a+5, inf)
-    :param x:
-    :return:
     """
     return nm.ones((nm.size(x), 1)) * 0
 
@@ -108,8 +94,6 @@ def const_u(x):
 def const_q(x):
     """
     piecewise constant (-inf, a],(a, a + 5](a+5, inf)
-    :param x:
-    :return:
     """
     return nm.ones(nm.size(x)) * .5
 
@@ -119,24 +103,24 @@ def const_q(x):
 #------------------------------#
 def ghump(x):
     """
-    :param x:
-    :return:
+    nice gaussian
     """
     return nm.exp(-200 * x**2)
 
 
 def gauss_init(x):
     """
-
-    :param x:
-    :return:
+    gaussian shifted to 0.2
     """
     return nm.array(ghump(x-.2))
 
 
 def gsmooth(x):
-    # return nm.piecewise(x, [x[:, 0] <= - 1, x[:, 0] >= -1, 1 < x[:, 0]],
-    #                     [0, lambda x: nm.exp(1 / (x ** 2 - 1)), 0])
+    """
+    .. :math: C_0^{\inf}
+    nm.piecewise(x, [x[:, 0] <= - 1, x[:, 0] >= -1, 1 < x[:, 0]],
+                         [0, lambda x: nm.exp(1 / (x ** 2 - 1)), 0])
+    """
     return  .3 * nm.piecewise(x, [x <= 0.1, x >= 0.1, .3 < x],
                                  [0, lambda x: nm.exp(1/((10*(x - .2))**2 - 1) + 1), 0])
 
@@ -144,8 +128,6 @@ def gsmooth(x):
 def superic(x):
     """
     All the initial conditions for the price of one, yayy!
-    :param x:
-    :return:
     """
     delta = 0.005
     beta = nm.log(4) / (36 * delta ** 2)
@@ -169,8 +151,6 @@ def superic(x):
 def sawtooth_q(x):
     """
     piecewise linear
-    :param x:
-    :return:
     """
     return signal.sawtooth(2 * nm.pi * 5 * x)
 
@@ -178,10 +158,6 @@ def sawtooth_q(x):
 #        Sinus and constant        #
 #----------------------------------#
 def cos_const_q(x):
-    """
-    :param x:
-    :return:
-    """
     return nm.piecewise(x, [x <= a + .3, a + .3 < x, a + .35 < x, a + .45 < x, a + .5 <= x],
                         [0, lambda t: (nm.cos(nm.pi * 20 * (t - .35)) + 1) / 2, 1,
                             lambda t: (nm.cos(nm.pi * 20 * (t - .45)) + 1) / 2, 0])
@@ -190,10 +166,6 @@ def cos_const_q(x):
 #   Quadratic and cubic function   #
 #----------------------------------#
 def quadr_cub(x):
-    """
-    :param x:
-    :return:
-    """
     return nm.piecewise(x, [x <= a + .1, a + .1 < x, a + .3 < x, a + .4 <= x],
                         [0, lambda t: -25*(t - 0.1) * (t - 0.5),
                             lambda t: -250*(t - 0.1) * (t - 0.1) * (t-0.4), 0])
@@ -205,8 +177,6 @@ def quadr_cub(x):
 def qsysinit(x):
     """
     initial values for nonconservative system
-    :param x:
-    :return:
     """
     return nm.array([ghump(x),
                      nm.zeros(len(x))]).swapaxes(0, 1)
@@ -215,8 +185,6 @@ def qsysinit(x):
 def wsysinit(x):
     """
     transformed innitial value for noncnons system solver
-    :param x:
-    :return:
     """
     return nm.stack((nm.sum(invRfunc(x)[:, 0, :] * gauss_init(x), 1),
                      nm.sum(invRfunc(x)[:, 1, :] * gauss_init(x), 1)), 1)
@@ -232,8 +200,6 @@ artb = 3.
 def Kfunc(x):
     """
     bulk modulus function
-    :param x:
-    :return:
     """
     return nm.piecewise(x, [x <= 0, x <= 5, 5 < x], [1, 1, 1])
 
@@ -241,8 +207,6 @@ def Kfunc(x):
 def rhofunc(x):
     """
     density function
-    :param x:
-    :return:
     """
     return nm.piecewise(x, [x <= 0, x > 0], [1, 4])
 
@@ -250,8 +214,6 @@ def rhofunc(x):
 def cfunc(x):
     """
     sound velocity function
-    :param x:
-    :return:
     """
     return nm.sqrt(Kfunc(x) / rhofunc(x))
 
@@ -259,8 +221,6 @@ def cfunc(x):
 def Zfunc(x):
     """
     impendance function
-    :param x:
-    :return:
     """
     return nm.sqrt(Kfunc(x) * rhofunc(x))
 
@@ -270,8 +230,6 @@ def Afunc(x):
     enviroment matrix
     A(x) =  [   0    , K(x) ]
             [1/rho(x),  0   ]
-    :param x:
-    :return:
     """
     return nm.array([[nm.zeros(len(x)), 1 / rhofunc(x)],
                      [Kfunc(x), nm.zeros(len(x))]]).swapaxes(0, 2)
@@ -281,8 +239,6 @@ def ATfunc(x):
     """
     transposed enviroment matrix,
     for adjoint conservative system
-    :param x:
-    :return:
     """
     return nm.transpose(Afunc(x), (0, 2, 1))
 
@@ -290,8 +246,6 @@ def ATfunc(x):
 def eigAfunc(x):
     """
     Jordan shape of A
-    :param x:
-    :return:
     """
     return nm.array([[-cfunc(x), nm.zeros(len(x))],
                      [nm.zeros(len(x)), cfunc(x)]]).swapaxes(0, 2)
@@ -300,8 +254,6 @@ def eigAfunc(x):
 def Rfunc(x):
     """
     eigen vectors of A
-    :param x:
-    :return:
     """
     return nm.array([[-Zfunc(x), nm.ones(len(x))],
                      [Zfunc(x), nm.ones(len(x))]]).swapaxes(0, 2)
@@ -310,8 +262,6 @@ def Rfunc(x):
 def invRfunc(x):
     """
     inverse egigen vectors of A
-    :param x:
-    :return:
     """
     return nm.array([[-1 / (2 * Zfunc(x)), nm.ones(len(x)) / 2],
                      [-1 / (2 * Zfunc(x)), nm.ones(len(x)) / 2]]).swapaxes(0, 2)
